Fix issues again from code review
